moved the factories to a namespace
